Decouple webapp from admin component as a lib.
