+++ .gitignore
@@ -2,3 +2,4 @@ npm-debug.log
 node_modules
 /build
 /dist
+/lib
