+++ package.json
@@ -4,7 +4,7 @@
   "description": "Kinto Web Administration Console in React.js",
   "scripts": {
     "build:gh-pages": "rimraf gh-pages && NODE_ENV=production KINTO_ADMIN_PUBLIC_PATH='/kinto-admin/' node_modules/.bin/webpack --progress --config webpack.prod.js && cp index.prod.html gh-pages/index.html",
-    "build:lib": "babel -d lib/ src/",
+    "build:lib": "rimraf lib && babel -d lib/ scripts/",
     "build:readme": "toctoc README.md -w",
     "flow-check": "node_modules/.bin/flow check",
     "lint": "node_modules/.bin/eslint scripts test test-browser",
